milestone overloads should stack on top of existing behavior
Currently if you defined a milestone for a before or after hook
it worked just fine since there was no existing function to call.
If, on the other hand, you wanted to add a milestone to a
pre-existing action, then your milestone would be called after the
existing function. This corrects that behavior.